backend/contactapi/serializers.py
from rest_framework import serializers

from contactapi.models import Contact


class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = '__all__'